Fix non-manifold export and 3MF colors for Bambu Studio
Root cause of non-manifold edges: quantized position dedup (Phase 2) ran
unconditionally, over-merging vertices that manifold-3d intentionally
kept separate at sharp edges. Collapsed thin triangles were removed by
the degenerate filter, creating holes. Fix: only run quantized dedup
when merge vectors are absent.

Extracted vertex dedup + degenerate filtering into shared meshClean.ts,
now used by both OBJ and 3MF exporters.

3MF color fix: Bambu Studio requires m:colorgroup/m:color from the
Materials Extension, not basematerials/base from the Core spec. Switched
to colorgroup format with 8-digit hex colors (#RRGGBBFF). Also applied
vertex deduplication and 6dp precision to 3MF export (was writing raw
split vertices from manifold-3d).
