+++ prompts/20260427-obj-color-export.md
@@ -31,4 +31,8 @@ OBJ files caused "3 non-manifold edges" warnings and slicing failures ("empty la
 
 4. **Float precision** — vertex coordinates written with 15+ decimal places (JavaScript float64 string representation of float32 values). Fix: round to 6 decimal places matching float32 precision and Blender convention.
 
-5. **usemtl groups split into separate shells** — Bambu Studio treats each OBJ `usemtl` material group as a separate part/shell that must be independently manifold. Since painted color regions are surface patches (not closed solids), every group had boundary edges at color boundaries. Fix: write all faces as one solid object without usemtl grouping. Note: even the user's reference Blender-exported colored_cube.obj has the same per-group boundary edge issue — this is a fundamental limitation of OBJ materials in slicers. For per-face colors in slicers, 3MF is the proper format.
+5. **usemtl groups split into separate shells** — Bambu Studio treats each OBJ `usemtl` material group as a separate part/shell that must be independently manifold. Since painted color regions are surface patches (not closed solids), every group had boundary edges at color boundaries. Fix: write all faces as one solid object without usemtl grouping.
+
+6. **Quantized dedup over-merged vertices** — Phase 2 quantized position dedup (1e-5 tolerance) ran unconditionally, even when authoritative merge vectors were present. This over-merged vertices that manifold-3d intentionally kept separate at sharp edges, collapsing thin triangles. The degenerate filter then removed them, creating holes → non-manifold edges. Fix: only run quantized dedup when merge vectors are absent. Extracted shared `meshClean.ts` module for both OBJ and 3MF exporters.
+
+7. **3MF used wrong color element for Bambu Studio** — Bambu Studio requires `<m:colorgroup>`/`<m:color>` from the 3MF Materials Extension, NOT `<basematerials>`/`<base>` from the Core spec. Also requires 8-digit hex with alpha (`#RRGGBBFF`). Fix: switched to colorgroup format. Also applied the same vertex dedup and 6dp precision to 3MF export.
