+++ src/export/obj.ts
@@ -1,78 +1,9 @@
 import type { MeshData } from '../geometry/types';
 import { downloadBlob, getExportFilename, getExportTitle } from './download';
 import { buildZip } from './zip';
+import { cleanMeshForExport } from './meshClean';
 
 const DEFAULT_COLOR = '#4a9eff';
-
-/**
- * Build a vertex remap table that merges duplicate vertices into canonical indices.
- *
- * Uses merge vectors from manifold-3d (authoritative) when available, otherwise
- * falls back to quantized position dedup (same tolerance as scadToManifold.ts).
- * Returns { remap, uniquePositions }.
- */
-function buildVertexRemap(meshData: MeshData) {
-  const { vertProperties, numVert, numProp, mergeFromVert, mergeToVert } = meshData;
-
-  // Union-find for vertex merging
-  const parent = new Uint32Array(numVert);
-  for (let i = 0; i < numVert; i++) parent[i] = i;
-
-  function find(x: number): number {
-    while (parent[x] !== x) { parent[x] = parent[parent[x]]; x = parent[x]; }
-    return x;
-  }
-  function union(a: number, b: number) {
-    const ra = find(a), rb = find(b);
-    if (ra !== rb) parent[ra] = rb;
-  }
-
-  // Phase 1: merge vectors from manifold-3d (exact pairs)
-  if (mergeFromVert && mergeToVert && mergeFromVert.length === mergeToVert.length) {
-    for (let i = 0; i < mergeFromVert.length; i++) {
-      union(mergeFromVert[i], mergeToVert[i]);
-    }
-  }
-
-  // Phase 2: quantized position dedup as fallback (catches anything merge vectors missed,
-  // or handles meshes without merge vectors like raw STL imports)
-  const quantize = (v: number) => Math.round(v * 1e5);
-  const posMap = new Map<string, number>();
-  for (let i = 0; i < numVert; i++) {
-    const x = quantize(vertProperties[i * numProp]);
-    const y = quantize(vertProperties[i * numProp + 1]);
-    const z = quantize(vertProperties[i * numProp + 2]);
-    const key = `${x},${y},${z}`;
-    const existing = posMap.get(key);
-    if (existing !== undefined) {
-      union(i, existing);
-    } else {
-      posMap.set(key, i);
-    }
-  }
-
-  // Flatten: assign sequential indices to unique roots
-  const rootToIndex = new Map<number, number>();
-  const uniquePositions: number[] = [];
-  const remap = new Uint32Array(numVert);
-
-  for (let i = 0; i < numVert; i++) {
-    const root = find(i);
-    let idx = rootToIndex.get(root);
-    if (idx === undefined) {
-      idx = uniquePositions.length / 3;
-      rootToIndex.set(root, idx);
-      uniquePositions.push(
-        vertProperties[root * numProp],
-        vertProperties[root * numProp + 1],
-        vertProperties[root * numProp + 2],
-      );
-    }
-    remap[i] = idx;
-  }
-
-  return { remap, uniquePositions };
-}
 
 /** Round a float to 6 decimal places (float32 has ~7 significant digits). */
 function f6(v: number): string {
@@ -80,21 +11,10 @@ function f6(v: number): string {
 }
 
 export function exportOBJ(meshData: MeshData, customName?: string): void {
-  const { triVerts, numTri, triColors } = meshData;
+  const { triVerts, triColors } = meshData;
   const title = getExportTitle();
 
-  const { remap, uniquePositions } = buildVertexRemap(meshData);
-
-  // Build non-degenerate triangle list (filter triangles that collapsed during merge)
-  const validTris: number[] = [];
-  for (let t = 0; t < numTri; t++) {
-    const a = remap[triVerts[t * 3]];
-    const b = remap[triVerts[t * 3 + 1]];
-    const c = remap[triVerts[t * 3 + 2]];
-    if (a !== b && b !== c && a !== c) {
-      validTris.push(t);
-    }
-  }
+  const { remap, uniquePositions, validTris } = cleanMeshForExport(meshData);
 
   // Detect whether we have painted triangles
   let hasColors = false;
