+++ src/export/threemf.ts
@@ -2,33 +2,36 @@ import type { MeshData } from '../geometry/types';
 import { get3MFUnitString } from '../geometry/units';
 import { downloadBlob, getExportFilename, getExportTitle } from './download';
 import { buildZip } from './zip';
+import { cleanMeshForExport } from './meshClean';
 
 export function export3MF(meshData: MeshData, customName?: string): void {
-  const { vertProperties, triVerts, numVert, numTri, numProp, triColors } = meshData;
+  const { triVerts, triColors } = meshData;
 
-  // Build vertices XML
+  // Deduplicate vertices and filter degenerate triangles (same as OBJ export)
+  const { remap, uniquePositions, validTris } = cleanMeshForExport(meshData);
+
+  // Build vertices XML (deduplicated, 6dp precision)
+  const numUniqueVerts = uniquePositions.length / 3;
   const vertices: string[] = [];
-  for (let i = 0; i < numVert; i++) {
-    const x = vertProperties[i * numProp];
-    const y = vertProperties[i * numProp + 1];
-    const z = vertProperties[i * numProp + 2];
+  for (let i = 0; i < numUniqueVerts; i++) {
+    const x = uniquePositions[i * 3].toFixed(6);
+    const y = uniquePositions[i * 3 + 1].toFixed(6);
+    const z = uniquePositions[i * 3 + 2].toFixed(6);
     vertices.push(`          <vertex x="${x}" y="${y}" z="${z}" />`);
   }
 
   // Collect distinct colors for basematerials (if triColors present)
-  // Index 0 is always the default base color for unpainted triangles
   const DEFAULT_COLOR = '#4a9eff';
-  const colorMap = new Map<string, number>(); // hex -> material index
+  const colorMap = new Map<string, number>();
   const materialColors: string[] = [];
   let hasColors = false;
 
   if (triColors) {
-    // Reserve index 0 for the default color
     colorMap.set(DEFAULT_COLOR, 0);
     materialColors.push(DEFAULT_COLOR);
 
     const painted = (triColors as Uint8Array & { _painted?: Uint8Array })._painted;
-    for (let t = 0; t < numTri; t++) {
+    for (const t of validTris) {
       const isPainted = painted ? painted[t] === 1 : (triColors[t * 3] !== 0 || triColors[t * 3 + 1] !== 0 || triColors[t * 3 + 2] !== 0);
       if (!isPainted) continue;
 
@@ -44,12 +47,12 @@ export function export3MF(meshData: MeshData, customName?: string): void {
     }
   }
 
-  // Build triangles XML — when colors exist, every triangle gets a pid
+  // Build triangles XML (using remapped vertex indices, filtered for degenerates)
   const triangles: string[] = [];
-  for (let t = 0; t < numTri; t++) {
-    const v1 = triVerts[t * 3];
-    const v2 = triVerts[t * 3 + 1];
-    const v3 = triVerts[t * 3 + 2];
+  for (const t of validTris) {
+    const v1 = remap[triVerts[t * 3]];
+    const v2 = remap[triVerts[t * 3 + 1]];
+    const v3 = remap[triVerts[t * 3 + 2]];
 
     if (hasColors && triColors) {
       const painted = (triColors as Uint8Array & { _painted?: Uint8Array })._painted;
@@ -62,7 +65,6 @@ export function export3MF(meshData: MeshData, customName?: string): void {
         const matIdx = colorMap.get(hex)!;
         triangles.push(`          <triangle v1="${v1}" v2="${v2}" v3="${v3}" pid="2" p1="${matIdx}" />`);
       } else {
-        // Unpainted triangles get the default base color
         triangles.push(`          <triangle v1="${v1}" v2="${v2}" v3="${v3}" pid="2" p1="0" />`);
       }
     } else {
@@ -70,16 +72,16 @@ export function export3MF(meshData: MeshData, customName?: string): void {
     }
   }
 
-  // Build basematerials XML block
-  let basematerialsXml = '';
+  // Build colorgroup XML block (Bambu Studio requires m:colorgroup, not basematerials)
+  let colorgroupXml = '';
   if (hasColors) {
-    const bases = materialColors.map((hex, i) =>
-      `      <base name="${i === 0 ? 'Default' : 'Color ' + i}" displaycolor="${hex}" />`
+    const colors = materialColors.map(hex =>
+      `      <m:color color="${hex.toUpperCase()}FF" />`
     ).join('\n');
-    basematerialsXml = `
-    <basematerials id="2">
-${bases}
-    </basematerials>`;
+    colorgroupXml = `
+    <m:colorgroup id="2">
+${colors}
+    </m:colorgroup>`;
   }
 
   // Escape XML special chars in title
@@ -92,7 +94,7 @@ ${bases}
 <model unit="${get3MFUnitString()}" xml:lang="en-US" xmlns="http://schemas.microsoft.com/3dmanufacturing/core/2015/02"${nsAttr}>
   <metadata name="Title">${title}</metadata>
   <metadata name="Application">Partwright</metadata>
-  <resources>${basematerialsXml}
+  <resources>${colorgroupXml}
     <object id="1" type="model"${hasColors ? ' pid="2" pindex="0"' : ''}>
       <mesh>
         <vertices>
